Optimize day17-1 (by ~77%) by changing input parse
The input parsing for day 17 is now done by directly looping over the
bytes in the input and grabbing the numbers that way. This is quicker
and therefore definitely also better.

This shaves off ~1µs, getting 17-1 to a sub-microsecond runtime (on my
machine anyway), which I thought was neat (and apparantly worth spending
my time on).

Technically this affects 17-2 as well, but the 1µs saved is neglible on
the ~400µs runtime.

(The timesaves of the unsafe unwraps aren't needed to get the desired
timesave but they do save some additional nanoseconds. Plus, it seemed
appropriate.)