Add `example_input` annotation to _derive
This is used to annotate a definition of an example input and it
replaces the macro_rules `example!` macro.

It will process the string as a typical heredoc (strip a single leading
& trailing newline, remove excess indentation) so that it is presented
to the program as it was on the AoC site (similar to a heredoc).

It also allows specifying the expected results, which will be stored
alongside the input in a single struct. In addition it can generate
tests validating that the actual result of processing the input matches
the expected result.