Ignore inputs directory
Advent of Code specifically requests you do _not_ include the inputs in
the repository, and I will respect this. The only reason they were
included in the past is that I was not aware of this request.

They don't mention including your answers so it'd probably be okay to
keep those, but that doesn't feel very useful without the input, so I've
chosen to exclude those as well.

These files have also been removed from the histories of both this
repository as well as the branches/repositories for the previous years.