Add workaround for inject_binary in rust analyzer
Rust analyzer presents the source of the files as a real file with an
empty path. This proc macro would attempt to read this empty path and
then panic while doing so.

As a workaround we now treat this source (real + empty path) specially,
generating a dummy `Bin` for it. This would be useless at runtime, but
it's fine for type checking/rust analyzer purposes.