Ignore binary tests in lib
The binaries being injected as modules in the main aoc bin (so that the
runner can access them) causes their tests to also be ran from this
location.

The result of this is that attempting to run the tests for a single day
will actually run the tests for all days (via the libs), and then the it
will run the tests you actually wanted an additional time (via the bin).
The same concept applies when running all tests, with all binary tests
running twice.