Improve error handling of parse! macro
Previously most statements would result in `Option`s which would then be
unwrapped at some point, which results in a generic panic which gives
little indication of what actually happened.

Instead these are now kept as/converted into `Result`s with a relevant
error message that provides some insight into what action failed.

In addition if the last token in the parse format is a literal it was
previously possible there was some input past this that would not be
considered, this will now result in an error.