Show calendars in web UI instead of a list
For every year in which I have at least some solutions a calendar of the
month of december is shown, with the background of days 1-25 colored
based on completion (0 stars = default, 1 star = silver, 2 stars =
gold). The days past the 25th are hidden.

Clicking a day opens up a screen with the elements that were previously
in the list (input, solve, examples, links, etc).