Include std/core & direct deps in generated docs
By default the generated docs will include all dependencies, but for
usage purposed I'm really only interested in dependencies I've added
myself, rather than transient dependencies, so I've limited it to only
include these.

In addition the standard library is _not_ included by default. This
means that any references to `std` will link to the official docs, and
that the search index won't include anything from `std`. The first point
is only a minor annoyance (you'll end up on a site where the rest of the
available stuff isn't in the index), but the second point is really
annoying as you will need to look at two pages to really get a sense of
what is available to you. I don't want to have to check two pages to
find out what I can do to an iterator, I just want it all in one place!

Recently a feature was added to rustdoc to generate the docs for
multiple crates separately and then combine the result. We can do this
for this repo + the standard library + direct dependencies and end up
with a single documentation including all those elements.