Make gauss_jordan_elimination easier to use
It now works on a `Matrix` of any numeric type (or really any type that
can be converted to `f64` according to the traits of the `num` crate).
It no longer mutates the matrix, instead working on a copy and returning
only the last column (once the solve is complete the rest of the matrix
is no longer useful).