Fix type error in 22-11
This only happens occasionally, as depending on enabled features there
is another type that can be compared to `u64` and it's unclear to rust
which type the empty vectors are of. Kinda silly as it wouldn't really
matter since they are empty, but of course the compiler can't know that.