Move utils to root, add prelude, simple bin setup
With most of the other code having moved to `aoc_runner` it no longer
makes sense to keep the `utils` mod, as that is pretty much all that is
left in the `aoc` crate's lib, so move the contents of this mod to the
crate root.

Add a prelude module for commonly-used stuff. This includes the
`parse!` macro & any extension methods.

Add a `setup!` macro that imports the prelude & calls the
`aoc_runner::single::generate_main!` macro. Every solution bin now just
calls this at the top, and that's all the boilerplate that's needed.
Nice and simple.

Made some minor changes to the template & `aoc_runner_derive` so that
there's no longer a bunch of commented-out stuff in the template, which
makes it slightly easier to get started with a new day.