runner: Use Solver struct in place of Option<fn>
Previously `Solver` was internal, and the `Bin`s would be generated with
an `Option<fn>` instead which would then be converted to a `Solver` when
needed. Now the `Solver` struct is moved to the `derive` mod (where
`Bin` also resides), and it is used directly in `Bin` and the generated
code.