Switch from cargo-udeps to cargo-machete
`cargo-udeps` requires building the entire project and is dreadfully
slow because of this. This makes it annoying as a pre-commit hook as I
have to sit there waiting for this step for 20+ second knowing I've not
touched the dependencies.

`cargo-machete` is _much_ faster, and for our purposes _more_ accurate
than `cargo-udeps` (though this is not universally true).