Update katex, fix fonts
Previously the katex CSS & JS were inlined in `katex.html`, but the
fonts were not, and the (relative) paths for the fonts did not work,
causing this to fallback to default fonts.

This changes it to load these files from a CDN, which means that these
font paths also point to this CDN and work correctly. It also updates it
to the newest version.