Show stacktrace in web UI on panic
The base crate that sets up the panic handler will print the stacktrace
to the console but will still return a generic error. There is a fork
that changes this to also return the stacktrace, which allows us to show
this in the UI.

Note also that the original crate has been archived, so it's not like we
risk losing out on updates by switching to a fork.