Reset WASM worker in web UI after a panic
Panic handling in the WASM build uses abort rather than unwind and
leaves the worked in an undefined state. Previously this wasn't very
noticable as this undefined state still works pretty well, but that was
just luck, it could easily have broken at any point. Now with the
improved error reporting this is more visible as this does break, which
means that only the first error will be reported properly, and any
subsequent runs will have the generic error again.

This introduces an additional class that wraps the worker & which
completely reloads it after a panic. Thankfully we already didn't pass
around any of the objects returned by the worker directly, so there are
no stale references that we need to be concerned with.