Add match operation to split parse macro
At its core this is a shorthand for `with (|item| match item { ... })`,
but it has some features beyond that.

The first feature is that the expression part of the match arm is
optional, and if not provided it defaults to passing through the matched
value unchanged. This can be useful when the `as $type` expression
(which can precede it) has already handled these cases.

The second feature is that the index at which a match occurs can be
captured in a separate variable. This works regardless of whether the
`indexed` flag is present. This is _mostly_ intended to be used in
conjunction for grids, where sometimes there are points of interest that
must be captured (e.g., the start or end point), but it can be used for
any kind of list.

The last feature is that there is a default fallback branch that will
either panic (for `match`) or return `None` (for `try match`).