Allow match operation in more contexts in parse
Previously it was only allowed as option to `split` or `cells`, but now
it is also allowed in a regular capture. This is mostly meant for use in
a nested capture in one of the aforementioned contexts, but it's
technically valid (though likely not useful) at the top level as well.