Remove visualization code
This hadn't been touched in a while and no longer compiles. In addition
I was never quite satisfied with how this worked and want to rethink the
base design to be easier to use (at some point). The old implementation
also predates the WASM builds, which will need to be taken into
consideration for any future replacement.