Avoid ok_or_else().unwrap() in parse! macro
This results in an error messages in the format `called
`Result::unwrap()` on an `Err` value: <err>`, where `<err>` is the
message returned by `ok_or_else`. This doesn't just add cruft, it also
adds a level of escaping to the error message which makes messages
containing anything that needs escaping much harder to read.