Improve build feedback of make targets
By default cargo's build output is pretty verbose, logging a line for
every built dependency. This is annoying for make targets that do two
things in a row (e.g., test + run) as this means that the output of the
first action will be scrolled off-screen by the building of the second
one.

Until now we had worked around that by passing the `--quiet` flag, but
that means we have no insights into the progress during the build step,
which causes the commands to appear to just hang for a while.

This adds a script which processes the cargo output to strip out these
"Compiled <crate-name>" lines while preserving the rest of the output.
This includes colors and escape commands, meaning that the progress bars
will work & update correctly.