Simplify point ranges
Previously these internally could use any of the range types in stdlib.
This was flexible, but it also made for a somewhat unwieldy type
definition which made using them annoying.

Now it will internally always use a plain `Range`. It can still be
created from any type of range, but it will convert this to a regular
range. Note that this is not always a lossless conversion; namely ranges
that would previously include the maximum value for that numeric type
will now exclude it, and inclusive upper bounds for non-integers may
include slightly more than before, but for our purposes neither of these
matter.