Overhaul puzzle runner structs
This adds a more well-defined hierarchy as another step towards non-AoC
puzzles. There is now the top-level `Series` (of which AoC is currently
the only one), `Chapter` (which replaces `Bin`, removing AoC-specific
items & assumptions), and `Part` (which replaces `Solver`).

The assumption about 2 parts existing is now gone, and each chapter can
contain any number of parts simply by implementing `partN` in the
chapter.

The WASM build is broken at this point, that'll follow in a separate
commit.