Skip puzzle_wasm in tests
This crate is a minimal wrapper around `puzzle_runner` and the crates
that implement puzzles. Because of this it really only contains WASM/web
specific things, which can't really be tested from the rust side, so
this crate has no tests.

It is quite expensive to compile though, so by skipping it we speed up
building the tests.