Speed up builds of chapter binaries
The lib of the puzzle crates contains the `Series`, which in turn
includes all `Chapter`s. This is required for the WASM builds as these
can only import from the lib of the crate. However, because each of the
binaries in this crate has access to its lib this means that the code
for _all_ chapters is now included in the build of every chapter, which
slows down builds quite a bit.

This adds a feature to `puzzle_runner_derive` that toggles whether the
`Series` should actually include the chapters like this (defaulting to
an empty list if this feature is disabled). This feature is enabled for
WASM builds, but not when directly running a single binary, resulting in
a substantial speedup of these builds.