Eliminate the 'multi' feature flag
This splits the `register_series` macro into two, with one
(`include_chapters`) finding & including the chapters, and the other
(`register_series`) setting up the `Series` object. The latter will
automatically use the former to include the chapters in the generated
object if the `include-chapters` flag is enabled. The former is used in
the multi & bench entrypoints to have access to the list of chapters
regardless of whether this flag is enabled.