Add workaround for inputs not being fetched
Somehow despite being marked as a dependency of the run target the input
files aren't actually downloaded, so we'll explicitly invoke make again
for now. This shouldn't be required and will likely break the `run-aoc`
target (as would the "intended" working in fairness), so this'll need
more work later, but this will do for now.