Speed up benchmark builds
The benchmark builds were slow for two reasons:

- All binaries were built beforing running them.
- They included all chapters instead of just the one being benchmarked.

The first one is a consequence of them being in the benchmarks folder;
when a benchmark from this location is ran all binaries are always built
first, and there is no way to change this. We avoid this by moving it to
the bin folder and running with `--bin {name` instead of `--bench
{name}`.

The second one is because the `bench` entrypoint was based on the
`multi` entrypoint. The reason for this is we need to disable the
regular benchmarking harness (because we use criterion), which is done
in `Cargo.toml`. There is no way to do this globally for all
benchmarks/binaries, only per target, which means that if we were to
have a separate benchmark entrypoint per-binary that would have to be in
the file. We could generate this, but that'd still result in diffs to
files tracked in the repo which we'd like to avoid.

The way we solve this is by making the code generated by the
`register_chapter` macro use the benchmark entrypoint if the `bench`
feature is enabled. We then have a single bin (`bench`) which we
configure not to use the harness in the config. In the makefile we write
the path of the to-be-benchmarked file to a file (which is not tracked
in the repo) and then benchmark with `--bin bench`. Finally, the `bench`
bin reads the file created by the makefile and includes the path
therein. This way running the benchmark target will build & benchmark
only that binary without making changes to any files tracked in the
repo.