Don't rerun for the *-set-baseline-* targets
Previously these targets would perform a new benchmark/profile and store
it as 'baseline'. This leads to a slightly awkward workflow when
attempting to do iterative improvements:

1. Set baseline at start.
2. Make changes.
3. Run non-baseline to test changes.
4. See improvement.
5. Set baseline (running the same code again).
6. Go back to 2.

This changes those targets to instead use the last result ('current')
and set it as the new baseline, which will instead give the following
workflow:

1. Run non-baseline at start.
2. Set result as baseline.
3. Make changes.
4. Run non-baseline to test changes.
5. See improvement.
6. Go back to 2.

This has an extra step at the start (which is pretty much instant), but
then avoids needlessly re-running with no code changes when changing the
baseline.