Partially automate saving expected results
When a part finished successfully and it doesn't yet have an expected
result saved, the result will automatically be written to the path that
the expected result should be at, with `.pending` appended. It also
adds a make target (`confirm`) that looks for these files and removes
this suffix. This way if the solution is correct I can simply run `make
confirm` to save it.