Add support for static extra arg to parts
Is it relatively common for AoC to have differences between the examples
and the real input that are not part of the input. For example, the
example input may have the described operations applied 100 times,
whereas the real input needs to have it applied 10000000 times.

This makes it possible to let the entrypoint for the part accept a
second argument to pass in such information. The `register_part`
attribute now accepts an argument (`arg`) which provides the value for
this second argument for the real input, while the examples accept
arguments in the form of `partN::arg` to set the values to use for the
second argument when running the example. For example:

```rust
fn part1(input: &str, cycles: usize) -> usize {
   ...
}

mod tests {
    #[example_input(part1 = 100, part1::arg = 1_000)]
    static EXAMPLE_INPUT: &str = "
      ...
    ";
}
```

In this example the `cycles` argument would be `1_000` when running the
examples and `1_000_000` when running with the real input.