Create CLI for controllers
This serves two purposes. The first (perhaps somewhat obvious) is to
make it possible to invoke these actions directly. This isn't super
useful for now since the input download was already automatically done
when needed, but it'll become more useful as more features are added to
controllers (e.g., determining the URL for a chapter).

The second is to avoid having to build the controller (& its
dependencies) for every bin. The controllers are (usually) going to
include something like an HTTP client & TLS handling, which are pretty
expensive to build compared to the rest of the code. In fact, the
addition of the current controller has roughly doubled the compile time
of the (release) builds for the AoC chapters.

Of course, we still want it to be available at runtime (to preserve the
input downloading automation & facilitate future automations), which is
no longer the case after this change. The next commit will fix this by
providing an implementation of `Controller` which wraps the controller
binary, allowing us to use it in the entrypoint for the chapter binaries
without having to compile it for each of these.