Create controller struct -> CLI proxy
This creates a struct which implements the `Controller` interface by
calling the (prebuilt) `controller.rs` binary in the same crate. The
`register_series` macro no longer takes a controller name, but instead
it just accepts a flag indicating that this setup should be used.

This allows us to use the functions implemented in the controller
interface without having to bundle it (& its dependencies) in each
chapter binary (which would bloat both filesize as well as compile
time).