Add validating results to controller
This only actually submits the answer if required; if the expected
result is already known it uses that information, and if an answer has
already been submitted & found to be invalid it will not be submitted
again.

It isn't automatically called after solving a part (yet) as I've not
figured out whether I want to do that; I'll keep an eye out to see how
frequently doing so would cause me to submit something I did not mean to
before deciding on that.

For now the answer continue to be written to a pending file, and there
is a new make target (`make submit`) which will submit these.