Use ehttp (instead of ureq) for AoC controller
This is done for WASM support. I don't intend to use any of the current
features of the controller in the web version (I have no desire to turn
this into a one-click solve-this-for-me thing), but I do intend to add
more features to the controller, some of which will be useful in the web
version.