Only use binary for HTTP instead of full controller
Previously it was set up in such a way that the real controllers were
defined in a binary, and any calls to the controller from the runner
would end up calling this binary. This was done to avoid pulling in the
HTTP stack into every binary, and it succeeded at that, but it does add
some overhead & a bunch of boilerplate.

I want to add additional features to the controller that (mostly) won't
actually require an HTTP stack & which might end up being called many
times, and having this overhead in the stack makes that somewhat
painful.

This change adds an interface between the controller & the HTTP stack,
and then moves the use of a precompiled binary to that layer. This means
that only the parts of the controller that require the HTTP stack end up
invoking a precombined binary while the rest is available directly
without additional overhead.