Optimize 15-20
This is done by only considering numbers that have a number of the
small primes as factors. A large portion of all integers have on or more
of these primes as factors, and as such any house number which does not
have these as factors will automatically miss out on a sufficiently
large number of elves that it's not worth considering.